pyproject.toml
[project]
name = "rss-paper-digest"
version = "0.1.0"
description = "RSSフィードから論文を取得し、Claude APIで関連度スコアリング・日本語要約を行うツール"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0",
    "anthropic>=0.50",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]
